src/images/linkedin-extractor/requirements.txt
aiohttp>=3.8.3
aiolimiter==1.0.0
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
boto3==1.26.52
botocore==1.29.52
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
jmespath==1.0.1
loguru==0.7.2
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.2
pandera==0.13.4
pyarrow==10.0.1
pydantic==1.10.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.31.0
s3transfer==0.6.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
python-dotenv==0.21.0 
